Resolves #1: Update to add delimiter parameter due to parameter changes in Vault search
In 2020, the SolidWorks vault search parameters where updated. This required
the use of the OR (|) operator when searching multiple filenames.
This commit updates the Base.parseText to accept a delimiter parameter,
with ' | ' as the default.

Updated non-vault related searches (TCM and comma) to use the delimiter=', '

Reference https://files.solidworks.com/Supportfiles/Whats_new/2020/English/whatsnew.pdf#page=163&zoom=100,86,545